Reorganized Files and baseclasses
