+++ vocab.py
@@ -1,6 +1,7 @@
 import os
 from pyparsing import *
 from pyparsing import common
+from .ottobase import OttoBase
 
 ### Keywords
 WHEN = CaselessKeyword('WHEN')
@@ -15,54 +16,15 @@ TRUE = CaselessKeyword('TRUE')
 CHANGES = CaselessKeyword('CHANGES')
 
 ### Classes
-class BaseVocab:
-    _parser = None
-    _value = None
-    _interpreter = None
-
-    def __init__(self,tokens):
-        self.tokens = tokens
-        for k,v in self.tokens.as_dict().items():
-            if type(v) == list and len(v) == 1:
-                setattr(self,k,v[0])
-            else:
-                setattr(self,k,v)
-
+class Vocab(OttoBase):
 
     def __str__(self):
         return f"{''.join(self.tokens)}"
 
-    @property
-    def interpreter(self):
-        return self._interpreter
-
-    @interpreter.setter
-    def interpreter(self, interpreter):
-        self._interpreter = interpreter
-
-    @property
-    def value(self):
-        return self._value
-
-    async def eval(self):
-        return self.value
-
-    @classmethod
-    def parser(cls):
-        return cls._parser.set_parse_action(cls)
-
-    @classmethod
-    def set_interpreter(cls, interpreter):
-        cls._interpreter = interpreter
-
-    @classmethod
-    def from_string(cls,string):
-        return cls.parser().parse_string(string)
-
-class StringValue(BaseVocab):
+class StringValue(Vocab):
     _parser = QuotedString("'",unquoteResults=True)("_value")
 
-class Numeric(BaseVocab):
+class Numeric(Vocab):
 
     _parser = common.number("_value")
 
@@ -70,29 +32,28 @@ class Numeric(BaseVocab):
         return f"{self._value}"
 
 
-class Hour(BaseVocab):
+class Hour(Vocab):
     _parser = Or(map(CaselessKeyword,["HOUR","HOUR"]))
 
     @property
     def as_seconds(self):
         return 3600
 
-class Minute(BaseVocab):
+class Minute(Vocab):
     _parser = Or(map(CaselessKeyword,["MINUTE","MINUTES"]))
 
     @property
     def as_seconds(self):
         return 60
 
-class Second(BaseVocab):
+class Second(Vocab):
     _parser = Or(map(CaselessKeyword,["SECOND","SECONDS"]))
 
     @property
     def as_seconds(self):
         return 1
 
-
-class TimeStamp(BaseVocab):
+class TimeStamp(Vocab):
     _digits = Combine(Char(nums) * 2).set_parse_action(lambda x: int(x[0]))
     _parser = _digits("_hours") + ":" + _digits("_minutes") + ":" +  _digits("_seconds")
 
@@ -100,23 +61,7 @@ class TimeStamp(BaseVocab):
     def as_seconds(self):
         return self._hours * 3600 + self._minutes * 60 + self._seconds
 
-
-
-class RelativeTime(BaseVocab):
-    _parser = Numeric.parser()("_count") + (Hour.parser() | Minute.parser() | Second.parser())("_unit")
-
-    @property
-    def as_seconds(self):
-        return self._count.value * self._unit.as_seconds
-
-    @property
-    def value(self):
-        return self.as_seconds
-
-term = Or([x.parser() for x in BaseVocab.__subclasses__()])
-
-
-class Entity(BaseVocab):
+class Entity(Vocab):
     _parser = common.identifier("_domain") + Literal(".") + common.identifier("_id") + Optional(":" + common.identifier("_attribute"))
 
     def __str__(self):
@@ -150,7 +95,7 @@ class Entity(BaseVocab):
             val = f"{self.domain}.{self._id}"
         return val
 
-class Var(BaseVocab):
+class Var(Vocab):
     _parser = Word('@',alphanums+'_')
 
     def __init__(self,tokens):
